% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{asymptotic_inference_internal}
\alias{asymptotic_inference_internal}
\title{Internal function: asymptotic inference}
\usage{
asymptotic_inference_internal(ss_back, PiHat, PhiHat, pairs, cormat, N)
}
\arguments{
\item{ss_back}{ss_back}

\item{PiHat}{PiHat}

\item{PhiHat}{PhiHat}

\item{pairs}{pairs}

\item{cormat}{cormat}

\item{N}{N}
}
\description{
Internal function: asymptotic inference
}
\examples{
\dontrun{

}
}
