% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{causal_inference}
\alias{causal_inference}
\title{Likelihood ratio tests}
\usage{
causal_inference(
  PiHat,
  PhiHat,
  ss_back,
  pairs,
  cormat,
  N,
  test_type = c("edge", "path")
)
}
\arguments{
\item{PiHat}{A pxp superset of the directed relations among phenotypes; output from superDAG function}

\item{PhiHat}{A qxp superset of the interventional relation; output from superDAG function}

\item{ss_back}{output from the summary_stats_preprocess function}

\item{pairs}{A list of vectors of length two to be tested, with each vector represents a directed node from the first to the second item of the vector}

\item{cormat}{pxp correlation matrix of phenotypes; this matrix can be estimated by empricial correlation of null SNP.}

\item{N}{A qxp matrix of corresponding sample size for each SNP-phenotype pair. A scalar is also acceptable when all SNP-phentype pairs have the same sample size}

\item{test_type}{"edge" refers to the test of multiple directed relations; "path" refers to the test of a directed pathway}
}
\value{
A list of likelihood ratio, and p-value
}
\description{
Likelihood ratio tests
}
\examples{
\dontrun{

}
}
