% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_superDAG}
\alias{estimate_superDAG}
\title{Estimating super-DAG from GWAS summary data}
\usage{
estimate_superDAG(
  beta_mat,
  se_mat,
  n,
  s_ = 0,
  maf_vec = NULL,
  geno_ref = NULL,
  thresh = 0.05,
  Xdata
)
}
\arguments{
\item{beta_mat}{A q by p matrix.
Regression coefficient of q phenptypes on p genotypes.}

\item{se_mat}{A q by p matrix.
The standard errors of beta_mat.}

\item{n}{A q by p matrix or a scalar.
Sample sizes for corresponding genotype-phenotype pairs.
A scalar indicates all genotype-phentype pairs have the same sample size.}

\item{s_}{A nonnegative scalar.
Regularize the genetic correlation matrix. The default is 0.}

\item{maf_vec}{A q vector of MAF for the genotypes.
Can be skipped if a genotype covariance matrix is given.}

\item{geno_ref}{A q by q genotype covariance matrix.}

\item{thresh}{A scalar. The threshold used in peeling procedure.}

\item{Xdata}{#TODO Genotype matrix in reference panel.}
}
\value{
A list of three quantities:
A p by p matrix an_mat, a q by p matrix iv_mat, and a scalar n_test
}
\description{
Estimating super-DAG from GWAS summary data
}
\examples{
\dontrun{
beta_mat <- matrix(rnorm(33 * 23), 33, 23)
se_mat <- matrix(rnorm(33 * 23)^2, 33, 23)
n <- 3393
Xdata <- matrix(rnorm(500 * 33), 500, 33)
estimate_superDAG(
    beta_mat = beta_mat,
    se_mat = se_mat, n = n, s_ = 0,
    geno_ref = var(Xdata), Xdata = Xdata
)
}
}
