% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_V}
\alias{estimate_V}
\title{(Internal function) Estimating V matrix based on summary data}
\usage{
estimate_V(beta_mat, se_mat, n, stats_list)
}
\arguments{
\item{beta_mat}{A q by p matrix.
Regression coefficient of q phenptypes on p genotypes.}

\item{se_mat}{A q by p matrix.
The standard errors of beta_mat.}

\item{n}{A q by p matrix or a scalar.
Sample sizes for corresponding genotype-phenotype pairs.
A scalar indicates all genotype-phentype pairs have the same sample size.}

\item{stats_list}{A list of XX, yy_vec, Xy_mat, maf_vec.
This is the output of preprocess() function.}
}
\value{
A q by p matrix V.
}
\description{
(Internal function) Estimating V matrix based on summary data
}
\examples{
\dontrun{
beta_mat <- matrix(rnorm(33 * 23), 33, 23)
se_mat <- matrix(rnorm(33 * 23)^2, 33, 23)
n <- 3393
estimate_V(beta_mat, se_mat, preprocess(beta_mat, se_mat, n))
}
}
