% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimateV_sum}
\alias{estimateV_sum}
\title{Internal function: estimating V matrix based on summary statistic}
\usage{
estimateV_sum(fullbetas, fullBetaSD, N, ss_back)
}
\arguments{
\item{fullbetas}{A qxp matrix of estimated regression coefficient (beta) of outcome on each genotype, where q is the number of SNPs and p is the number of phenotype.}

\item{fullBetaSD}{A qxp matrix of the standard error of matrix 'fullbetas'}

\item{N}{A qxp matrix of corresponding sample size for each SNP-phenotype pair. A scalar is also acceptable when all SNP-phentype pairs have the same sample size}

\item{ss_back}{ss_back}
}
\description{
Internal function: estimating V matrix based on summary statistic
}
\examples{
\dontrun{

}
}
