% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{preprocess}
\alias{preprocess}
\title{Estimating key quantities from GWAS summary data}
\usage{
preprocess(beta_mat, se_mat, n, s_ = 0, maf_vec = NULL, geno_ref = NULL)
}
\arguments{
\item{beta_mat}{A q by p matrix.
Regression coefficient of q phenptypes on p genotypes.}

\item{se_mat}{A q by p matrix.
The standard errors of beta_mat.}

\item{n}{A q by p matrix or a scalar.
Sample sizes for corresponding genotype-phenotype pairs.
A scalar indicates all genotype-phentype pairs have the same sample size.}

\item{s_}{A nonnegative scalar.
Regularize the genetic correlation matrix. The default is 0.}

\item{maf_vec}{A q vector of MAF for the genotypes.
Can be skipped if a genotype covariance matrix is given.}

\item{geno_ref}{A q by q genotype covariance matrix.}
}
\value{
A list of quantities:
q by q matrix XX, p vector yy_vec, q by p matrix Xy_mat, q vector maf_vec
}
\description{
Estimating key quantities from GWAS summary data
}
\examples{
beta_mat <- matrix(rnorm(33 * 23), 33, 23)
se_mat <- matrix(rnorm(33 * 23)^2, 33, 23)
n <- 3393
Xdata <- matrix(rnorm(500 * 33), 500, 33)
preprocess(beta_mat, se_mat, n, geno_ref = var(Xdata))

}
