% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{pseudoBIC.internal}
\alias{pseudoBIC.internal}
\title{Internal function: pseudoBIC}
\usage{
pseudoBIC.internal(penalizedBetas, betas, ses, Ni, sds, xtx, ytyEst, xtyEst)
}
\arguments{
\item{penalizedBetas}{penalizedBetas}

\item{betas}{betas}

\item{ses}{ses}

\item{Ni}{Ni}

\item{sds}{sds}

\item{xtx}{xtx}

\item{ytyEst}{ytyEst}

\item{xtyEst}{xtyEst}
}
\description{
This function computes pseudoBIC.
}
\examples{
\dontrun{

}
}
