% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{pseudoBIC}
\alias{pseudoBIC}
\title{(Internal function) Computing pseudoBIC}
\usage{
pseudoBIC(beta_vec, n, XX, yy, Xy)
}
\arguments{
\item{beta_vec}{A q vector.
Coefficients of the joint regression of a phenotype over genotypes.}

\item{n}{A q vector of sample sizes.}

\item{XX}{A q by q genotype covariance matrix.}

\item{yy}{A scalar.}

\item{Xy}{A q vector.}
}
\value{
A list of three vectors:
AIC, BIC, SSE, and the estimated error variance.
}
\description{
(Internal function) Computing pseudoBIC
}
\examples{
\dontrun{
X <- matrix(rnorm(100 * 10), 100, 10)
y <- rnorm(100)
XX <- t(X) \%*\% X
Xy <- t(X) \%*\% y
yy <- t(y) \%*\% y
n <- 100
beta_vec <- matrix(rnorm(10 * 5), 10, 5)
pseudoBIC(beta_vec, n, XX, yy, Xy)
}
}
