% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{summary_stats_preprocess}
\alias{summary_stats_preprocess}
\title{Title Estimation of key quantities from GWAS summary statistics}
\usage{
summary_stats_preprocess(
  fullbetas,
  fullses,
  N,
  xtxdiag = 0,
  ref = FALSE,
  geno.ref.txt = NULL
)
}
\arguments{
\item{fullbetas}{A qxp matrix of estimated regression coefficient (beta) of outcome on each genotype, where q is the number of SNPs and p is the number of phenotype.}

\item{fullses}{A qxp matrix of the standard error of matrix 'fullbetas'}

\item{N}{A qxp matrix of corresponding sample size for each SNP-phenotype pair. A scalar is also acceptable when all SNP-phentype pairs have the same sample size}

\item{xtxdiag}{A constant number used to regularize the genetic correlation matrix; by default, no penalization is used.}

\item{ref}{ref = T if users provide a reference panel; ref = F if all SNPs are pruned and standardized and no reference panel is needed}

\item{geno.ref.txt}{The location of the user-provided reference panel with the genotype of the q SNPs; only txt files are acceptable; The order SNPs needs to be matched with fullbetas and fullses.}
}
\value{
A list of quantities: qxq matrix xtx, p-vector ytyVec, qxp matrix xtyMat, q vector sds
}
\description{
Title Estimation of key quantities from GWAS summary statistics
}
\examples{
\dontrun{
sim <- sim1(ES = 5, Utxt = "../simulations/Uhat.txt", Wtxt = "../simulations/What1.txt")
bfile <- "../simulations/chr22.26SNPs.ref.3000_3.txt"
samplesize <- 3000
Xdata <- input_data(samplesize.sim = samplesize, sim = sim, sigma2txt = "../simulations/sigma2.txt")
Y <- simulate_Y(Xdata$X.std, Xdata$sigma2, sim, prots_name = Xdata$prots)
cormat <- cor(t(Y))
ss <- cal_sum_stats(Xdata$X.std, Xdata$snp_names, Xdata$sim_snps_bim, Y)
fullBetas <- ss$fullBetas
fullBetaSD <- ss$fullBetaSD
N <- ss$fullBetaSS
ss_back <- summary_stats_preprocess(fullbetas = fullBetas, fullses = fullBetaSD, N = N, xtxdiag = 0, ref = TRUE, geno.ref.txt = bfile)
}
}
