% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{superDAG}
\alias{superDAG}
\title{Estimation of superDAG from GWAS summary statistics}
\usage{
superDAG(
  fullBeta,
  fullBetaSD,
  N,
  xtxdiag = 0,
  ref = F,
  geno.ref.text = NULL,
  thresh = 0.05,
  Xdata
)
}
\arguments{
\item{fullBeta}{A qxp matrix of estimated regression coefficient (beta) of outcome on each genotype, where q is the number of SNPs and p is the number of phenotype.}

\item{fullBetaSD}{A qxp matrix of the standard error of matrix 'fullbetas'}

\item{N}{A qxp matrix of corresponding sample size for each SNP-phenotype pair. A scalar is also acceptable when all SNP-phentype pairs have the same sample size}

\item{xtxdiag}{A constant number used to regularize the genetic correlation matrix; by default, no penalization is used.}

\item{ref}{ref = T if users provide a reference panel; ref = F if all SNPs are pruned and standardized and no reference panel is needed}

\item{geno.ref.text}{The location of the user-provided reference panel with the genotype of the q SNPs; only txt files are acceptable; The order SNPs needs to be matched with fullbetas and fullses.}

\item{thresh}{The threshold of V used in the peeling algorithm.}

\item{Xdata}{Xdata}
}
\value{
A list of three elements: PiHat, PhiHat, and effect.test
}
\description{
Estimation of superDAG from GWAS summary statistics
}
\examples{
\dontrun{

}
}
