% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{TLP_afterpeel_U}
\alias{TLP_afterpeel_U}
\title{TLP after peeling}
\usage{
TLP_afterpeel_U(PiHat, PhiHat, ss_back, cormat, N, penalizeXX = T)
}
\arguments{
\item{PiHat}{A pxp superset of the directed relations among phenotypes; output from superDAG function}

\item{PhiHat}{A qxp superset of the interventional relation; output from superDAG function}

\item{ss_back}{output from the summary_stats_preprocess function}

\item{cormat}{pxp correlation matrix of phenotypes; this matrix can be estimated by empricial correlation of null SNP.}

\item{N}{A qxp matrix of corresponding sample size for each SNP-phenotype pair. A scalar is also acceptable when all SNP-phentype pairs have the same sample size}

\item{penalizeXX}{penalizeXX = T if users prefer penalzing the IV matrix after peeling; penalizeXX = F if users do not.}
}
\value{
A list of two elements: W.new, U.new
}
\description{
TLP after peeling
}
\examples{
\dontrun{

}
}
