% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{TLP_U}
\alias{TLP_U}
\title{Truncated Lasso penalized estimation after peeling}
\usage{
TLP_U(an_mat, iv_mat, stats_list, corr_Y, n, if_penalize_iv = TRUE)
}
\arguments{
\item{an_mat}{A p by p ancestral relation matrix}

\item{iv_mat}{A q by p interventional relation matrix}

\item{stats_list}{A list output by preprocess() function}

\item{corr_Y}{A p by p correlation matrix of phenotypes;
Can be estimated by empricial correlation of null SNPs.}

\item{n}{A q by p matrix or a scalar.
Sample sizes for corresponding genotype-phenotype pairs.
A scalar indicates all genotype-phentype pairs have the same sample size.}

\item{if_penalize_iv}{Binary.
TRUE if users prefer penalzing the IVs; FALSE otherwise.}
}
\value{
A list of two elements:
A p by p adjacency matrix U, a q by p interventional matrix W
}
\description{
Truncated Lasso penalized estimation after peeling
}
\examples{
\dontrun{
beta_mat <- matrix(rnorm(33 * 23), 33, 23)
se_mat <- matrix(rnorm(33 * 23)^2, 33, 23)
n <- 3393
Xdata <- matrix(rnorm(500 * 33), 500, 33)
res <- estimate_superDAG(beta_mat, se_mat, n, Xdata)
an_mat <- res$an_mat
iv_mat <- res$iv_mat
y <- matrix(rnorm(100, 23), 100, 23)
corr_Y <- cor(y)
stats_list <- preprocess(beta_mat, se_mat, n)
TLP_U(an_mat, iv_mat, stats_list, corr_Y, n)
}
}
