% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_order.R
\name{topological_order}
\alias{topological_order}
\title{Internal function: topological order}
\usage{
topological_order(v, thresh = 0.15)
}
\arguments{
\item{v}{v matrix produced by peeling}

\item{thresh}{threshold}
}
\description{
This function implements peeling algorithm.
}
\examples{
\dontrun{

}
}
