% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{estimate_Sigma}
\alias{estimate_Sigma}
\title{Internal function: estimating Sigma using summary statistics in the regression of Y ~ X(IV) + Y(ancestor) on j's y}
\usage{
estimate_Sigma(output_prepare, PhiHatj, PiHatj, cormat, N, j, ss_back)
}
\arguments{
\item{output_prepare}{output_prepare}

\item{PhiHatj}{PhiHatj}

\item{PiHatj}{PiHatj}

\item{cormat}{cormat}

\item{N}{N}

\item{j}{j}

\item{ss_back}{ss_back}
}
\description{
Internal function: estimating Sigma using summary statistics in the regression of Y ~ X(IV) + Y(ancestor) on j's y
}
\examples{
\dontrun{

}
}
