% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{peeling}
\alias{peeling}
\title{Internal function: peeling}
\usage{
peeling(V, thresh)
}
\arguments{
\item{V}{V matrix produced by peeling algorithm}

\item{thresh}{threshold}
}
\description{
This function implements peeling algorithm.
}
\examples{
\dontrun{

}
}
