% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{topological_order}
\alias{topological_order}
\title{(Internal function) Peeling procedure implementation.}
\usage{
topological_order(v, thresh = 0.15)
}
\arguments{
\item{v}{A q by p matrix. V matrix.}

\item{thresh}{A positive scalar. Threshold.}
}
\description{
(Internal function) Peeling procedure implementation.
}
\examples{
\dontrun{
v <- matrix(rnorm(15 * 10), 15, 10)
thresh <- 0.8
topological_order(v, thresh)
}
}
